requirements.txt
reflex==0.5.*
sqlmodel
chromadb==0.5.*
pydantic
pypdf
pytesseract
Pillow
python-docx
openai
psycopg[binary]
supabase>=2.6.0
python-dotenv>=1.0.0
